clang-cross-canadian_git.bbappend: Add dependency on nativesdk-clang
meta-clang/.../clang-cross-canadian_git.bb has a DEPEND on nativesdk-clang
(and other packages), but no RDEPEND. Personally I think this is a bug,
and it should be a RDEPEND, but for the time being work round it by
adding an RDEPEND here.

Note in the Linux build nativesdk-clang gets pulled into the SDK through
the RDEPENDS added by nativesdk-packagegroup-sdk-host.bbappend but this
doesn't help with baremetal builds.