llvm-project-source.bbappend: Add patches for compiler driver
We're using the header files and run time libraries generated by the
gcc toolchain at the moment. The puts the header files and libraries
in .../usr/lib, so as long as we fix up the path, and don't pass
-nostdsysteminc to the compiler, it should just work. Similarly for the
run tme libraries, we want to default to libgcc.a rather than clang_rt,
so switch the default and don't try and use libunwind, as it appears
unwind functionality is already part of libgcc.a.